+++ src/super/i-block/modules/state/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2023-??-??)
+
+#### :bug: Bug Fix
+
+* Fixed resetting router state
+
 ## v3.54.0 (2023-07-28)
 
 #### :rocket: New Feature